+++ TaskManager/Models/Comment.cs
@@ -2,11 +2,11 @@
 {
     public class Comment
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Message { get; set; }
         public Users User { get; set; }
         public string UserName { get; set; }
-        public int TaskId { get; set; } 
+        public long TaskId { get; set; } 
         public Tasks Tasks { get; set; }
     }
 }
