+++ TaskManager/Services/Interfaces/ICommentService.cs
@@ -6,8 +6,8 @@ namespace TaskManager.Services.Interfaces
 {
     public interface ICommentService
     {
-        public Task<IBaseResponse<ICollection<Comment>>> GetByTask(int taskId);
-        public Task<IBaseResponse<Comment>> Remove(int id);
+        public Task<IBaseResponse<ICollection<Comment>>> GetByTask(long taskId);
+        public Task<IBaseResponse<Comment>> Remove(long id);
         public Task<IBaseResponse<Comment>> Create(CreateCommentVM comment);
     }
 }
