+++ TaskManager/Services/Interfaces/ITaskService.cs
@@ -9,12 +9,12 @@ namespace TaskManager.Services.Interfaces
     {
         public Task<IBaseResponse<GetTaskVM>> Create(CreateTaskVM task);
         public Task<IBaseResponse<ICollection<GetTaskVM>>> GetAll();
-        public Task<IBaseResponse<ICollection<GetTaskVM>>> GetAllDone(int themeId);
-        public Task<IBaseResponse<ICollection<GetTaskVM>>> GetAllNotDone(int themeId);
-        public Task<IBaseResponse<GetTaskVM>> GetById(int id);
-        public Task<IBaseResponse<GetTaskVM>> Remove(int id);
-        public Task<IBaseResponse<GetTaskVM>> Update(int id, UpdateTaskVM task);
-        public Task<IBaseResponse<GetTaskVM>> Complite(int id);
+        public Task<IBaseResponse<ICollection<GetTaskVM>>> GetAllDone(long themeId);
+        public Task<IBaseResponse<ICollection<GetTaskVM>>> GetAllNotDone(long themeId);
+        public Task<IBaseResponse<GetTaskVM>> GetById(long id);
+        public Task<IBaseResponse<GetTaskVM>> Remove(long id);
+        public Task<IBaseResponse<GetTaskVM>> Update(long id, UpdateTaskVM task);
+        public Task<IBaseResponse<GetTaskVM>> Complite(long id);
 
     }
 }
