+++ TaskManager/Services/Interfaces/IUserService.cs
@@ -9,9 +9,9 @@ namespace TaskManager.Services.Interfaces
         public Task<IBaseResponse<Users>> Register(AccountVM task);
         public Task<IBaseResponse<string>> LogIn(LogInVM task);
         public Task<IBaseResponse<ICollection<Users>>> GetAll();
-        public Task<IBaseResponse<Users>> Remove(int id);
-        public Task<IBaseResponse<Users>> ChangeRole(int id);
-        public Task<IBaseResponse<Users>> ChangePassword(int userId, ChangePasswordVM changePassword);
-        public Task<IBaseResponse<Users>> ChangeEmail(int userId, ChangeEmail changePassword);
+        public Task<IBaseResponse<Users>> Remove(long id);
+        public Task<IBaseResponse<Users>> ChangeRole(long id);
+        public Task<IBaseResponse<Users>> ChangePassword(long userId, ChangePasswordVM changePassword);
+        public Task<IBaseResponse<Users>> ChangeEmail(long userId, ChangeEmail changePassword);
     }
 }
