+++ TaskManager/ViewModels/Comments/CreateCommentVM.cs
@@ -2,7 +2,7 @@
 {
     public class CreateCommentVM
     {
-        public int TaskId { get; set; }
+        public long TaskId { get; set; }
         public string Message { get; set; }
         public string UserName { get; set; }
