+++ TaskManager/ViewModels/Tasks/UpdateTaskVM.cs
@@ -6,7 +6,7 @@
         public string TaskDescription { get; set; }
         public string Status { get; set; }
         public string Priority { get; set; }
-        public int ThemeId { get; set; }
+        public long ThemeId { get; set; }
         public string DeadLine { get; set; }
         public string Participates { get; set; }
     }
